Algumas correções
